+++ src/converters/TandemReader.cpp
@@ -306,7 +306,8 @@ void TandemReader::readSpectra(const tandem_ns::group &groupObj, bool isDecoy,
 	      spos = fileId.find('.');
 	      if (spos != std::string::npos) fileId.erase(spos);
 	      //Create id
-	      std::string psmId = createPsmId(fileId, parentIonMass, spectraId, charge, rank);
+	      std::string psmId = createPsmId(fileId, parentIonMass, 
+          static_cast<unsigned int>(spectraId), static_cast<int>(charge), static_cast<unsigned int>(rank));
       	createPSM(domain, parentIonMass, charge, sumI, maxI, isDecoy, database, peptideProteinMap, psmId, spectraId);
       	++rank;
       }//End of if rank<=po.hitsPerSpectrum
@@ -414,7 +415,7 @@ void TandemReader::createPSM(const tandem_ns::peptide::domain_type &domain,
     if (freqAA.find(peptideS[ix]) == string::npos) {
       int accession = ptmMap[peptideS[ix]];
       std::auto_ptr< percolatorInNs::uniMod > um_p(new percolatorInNs::uniMod(accession));
-      std::auto_ptr< percolatorInNs::modificationType > mod_p( new percolatorInNs::modificationType(ix));
+      std::auto_ptr< percolatorInNs::modificationType > mod_p( new percolatorInNs::modificationType(static_cast<int>(ix)));
       mod_p->uniMod(um_p);
       peptide_p->modification().push_back(mod_p);      
       peptideS.erase(ix--,1);
@@ -449,12 +450,12 @@ void TandemReader::createPSM(const tandem_ns::peptide::domain_type &domain,
   f_seq.push_back(hyperscore);
   f_seq.push_back(hyperscore - next_hyperscore);
   //ions fractions
-  if (a_score) f_seq.push_back(aions / peptide.size());
-  if (b_score) f_seq.push_back(bions / peptide.size());
-  if (c_score) f_seq.push_back(cions / peptide.size());
-  if (x_score) f_seq.push_back(xions / peptide.size());
-  if (y_score) f_seq.push_back(yions / peptide.size());
-  if (z_score) f_seq.push_back(zions / peptide.size());
+  if (a_score) f_seq.push_back(aions / static_cast<double>(peptide.size()));
+  if (b_score) f_seq.push_back(bions / static_cast<double>(peptide.size()));
+  if (c_score) f_seq.push_back(cions / static_cast<double>(peptide.size()));
+  if (x_score) f_seq.push_back(xions / static_cast<double>(peptide.size()));
+  if (y_score) f_seq.push_back(yions / static_cast<double>(peptide.size()));
+  if (z_score) f_seq.push_back(zions / static_cast<double>(peptide.size()));
   //Mass
   f_seq.push_back(parentIonMass);
   f_seq.push_back(mass_diff);
@@ -492,7 +493,7 @@ void TandemReader::createPSM(const tandem_ns::peptide::domain_type &domain,
   //Save the psm
   std::auto_ptr< percolatorInNs::peptideSpectrumMatch > psm_p(
       new percolatorInNs::peptideSpectrumMatch(features_p, peptide_p, psmId, 
-      isDecoy, parentIonMass, calculated_mass, charge));
+      isDecoy, parentIonMass, calculated_mass, static_cast<int>(charge)));
   
   std::vector<std::string>::const_iterator poIt;
   for (poIt = proteinOccurences.begin(); poIt != proteinOccurences.end(); ++poIt) {
@@ -500,7 +501,7 @@ void TandemReader::createPSM(const tandem_ns::peptide::domain_type &domain,
     psm_p->occurence().push_back(oc_p);
   }
   
-  database->savePsm(spectraId, psm_p);
+  database->savePsm(static_cast<unsigned int>(spectraId), psm_p);
 }
 
 void TandemReader::read(const std::string &fn, bool isDecoy,
